[utils] Fix bug in 0b9c08b47bb5e95c21b067044ace4e824d19a9c2
* Cache of `supports_terminal_sequences` must be reset after enabling VT mode
* and move `windows_enable_vt_mode` to utils to avoid cyclic imports
